txm: introduce tx_memory_manager
Since transaction can allocate objects on its own, we need a special memory manager
for transactions that will be a base class for all the memory managers for
transaction managers.

One should not create an instance of tx_memory_manager as an independent unit.
It should be used only as a base class.

Module tx_memory requires to register transaction with a special method before
using tx_memory methods with it.

tx_memory contains proxies for all the region and mempool [de]allocation methods
instead of region_truncate() - the only way to truncate txn's region is to clear it.

Part of #6150